+++ src/App.jsx
@@ -33,7 +33,7 @@ const App = () => (
 function Layout() {
   return (
     <div className="flex flex-col min-h-screen justify-between">
-      <header className="sticky top-0 z-50 sm:px-16 px-6 flex justify-center items-center bg-checkBG">
+      <header className="sticky top-0 z-50 sm:px-16 px-6 flex justify-center items-center bg-checkBG h-28">
         <div className="xl:max-w-[1280px] w-full">
           <Navbar />
         </div>
