+++ src/components/Navbar.jsx
@@ -41,7 +41,7 @@ const Navbar = () => {
   }
 
   return (
-    <nav className="w-full flex py-6 justify-between items-center navbar ">
+    <nav className="w-full flex py-6 justify-between items-center navbar">
       <ToastContainer />
       {/** Logo */}
       <Link to="/">
