+++ src/components/index.js
@@ -5,8 +5,6 @@ import Approach from "./Approach"
 import CTA from "./CTA"
 import Check from "./Check"
 import Faq from "./Faq"
-import Team from "./Team";
-import JoinUs from "./JoinUs";
 
 export {
   Navbar,
@@ -15,7 +13,5 @@ export {
   Approach,
   Check,
   CTA,
-  Faq,
-  Team,
-  JoinUs
+  Faq
 };
